+++ src/StreamDeck.SDK/PiModel.fs
@@ -5,33 +5,31 @@ open Browser.Types
 open Dto
 
 [<RequireQualifiedAccess>]
-type PiInEvent =
+type PiEvent =
     /// When `websocket.onopen` this event diliver initial state to property inspector.
-    | Connected of startArgs: StartArgs * replyAgent: MailboxProcessor<PiOutEvent>
-    /// Event received after calling the getSettings API to retrieve the persistent data stored for the action.
-    | DidReceiveSettings of event: Event * payload: ActionPayload
-    /// Event received after calling the getGlobalSettings API to retrieve the global persistent data.
+    | Connected of startArgs: StartArgs * replyAgent: MailboxProcessor<PiCommand>
+    /// Occurs when the plugin receives the global settings from the Stream Deck.
     | DidReceiveGlobalSettings of setting: obj
-    /// Event received by the Property Inspector when the plugin uses the `sendToPropertyInspector` event.
-    | SendToPropertyInspector of event: Event
+    /// Occurs when a message was received from the plugin.
+    | DidReceivePluginMessage of event: Event
+    /// Occurs when the settings associated with an action instance are requested, or when the the settings were updated by the property inspector.
+    | DidReceiveSettings of event: Event * payload: ActionPayload
 
-and [<RequireQualifiedAccess>] PiOutEvent =
-    /// Save data persistently for the action's instance.
-    | SetSettings of payload: obj
-    /// Request the persistent data for the action's instance.
-    | GetSettings
-    /// Save data securely and globally for the plugin.
-    | SetGlobalSettings of payload: obj
-    /// Request the global persistent data.
+and [<RequireQualifiedAccess>] PiCommand =
+    /// Gets the global settings associated with the plugin. Causes DidReceiveGlobalSettings to be emitted.
     | GetGlobalSettings
-    /// Open an URL in the default browser.
+    /// Gets the settings associated with an instance of an action. Causes DidReceiveSettings to be emitted.
+    | GetSettings
+    /// Opens the URL in the user's default browser.
     | OpenUrl of url: string
-    /// Write a debug log to the logs file.
-    | LogMessage of message: string
-    /// Send a payload to the plugin.
+    /// Sends a message to the plugin.
     | SendToPlugin of payload: obj
+    /// Sets the global settings associated with the plugin.
+    | SetGlobalSettings of payload: obj
+    /// Sets the settings associated with an instance of an action.
+    | SetSettings of payload: obj
 
-let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor<PiOutEvent> =
+let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor<PiCommand> =
     let inPropertyInspectorUUID = args.UUID
 
     MailboxProcessor.Start(fun inbox ->
@@ -44,47 +42,39 @@ let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor
                 console.log($"PI sent event %A{msg}", msg)
 
                 match msg with
-                | PiOutEvent.SetSettings payload ->
-                    sendJson {|
-                        event = "setSettings"
-                        // An opaque value identifying the Property Inspector. This value is received by the Property Inspector as parameter of the connectElgatoStreamDeckSocket function.
-                        context = inPropertyInspectorUUID
-                        payload = payload
-                    |}
-                | PiOutEvent.GetSettings ->
-                    sendJson {|
-                        event = "getSettings"
-                        // An opaque value identifying the Property Inspector. This value is received by the Property Inspector as parameter of the connectElgatoStreamDeckSocket function.
-                        context = inPropertyInspectorUUID
-                    |}
-                | PiOutEvent.SetGlobalSettings payload ->
-                    sendJson {|
-                        event = "setGlobalSettings"
-                        // An opaque value identifying the Property Inspector (inPropertyInspectorUUID). This value is received during the Registration procedure.
-                        context = inPropertyInspectorUUID
-                        payload = payload
-                    |}
-                | PiOutEvent.GetGlobalSettings ->
+                | PiCommand.GetGlobalSettings ->
                     sendJson {|
                         event = "getGlobalSettings"
-                        // An opaque value identifying the Property Inspector (inPropertyInspectorUUID). This value is received during the Registration procedure.
                         context = inPropertyInspectorUUID
                     |}
-                | PiOutEvent.OpenUrl url ->
+                | PiCommand.GetSettings ->
+                    sendJson {|
+                        action = args.ActionInfo.Value.action
+                        event = "getSettings"
+                        context = inPropertyInspectorUUID
+                    |}
+                | PiCommand.OpenUrl url ->
                     sendJson {|
                         event = "openUrl"
                         payload = {| url = url |}
                     |}
-                | PiOutEvent.LogMessage message ->
-                    sendJson {|
-                        event = "logMessage"
-                        payload = {| message = message |}
-                    |}
-                | PiOutEvent.SendToPlugin payload ->
+                | PiCommand.SendToPlugin payload ->
                     sendJson {|
                         action = args.ActionInfo.Value.action
                         event = "sendToPlugin"
-                        // An opaque value identifying the Property Inspector. This value is received by the Property Inspector as parameter of the connectElgatoStreamDeckSocket function.
+                        context = inPropertyInspectorUUID
+                        payload = payload
+                    |}
+                | PiCommand.SetGlobalSettings payload ->
+                    sendJson {|
+                        event = "setGlobalSettings"
+                        context = inPropertyInspectorUUID
+                        payload = payload
+                    |}
+                | PiCommand.SetSettings payload ->
+                    sendJson {|
+                        action = args.ActionInfo.Value.action
+                        event = "setSettings"
                         context = inPropertyInspectorUUID
                         payload = payload
                     |}
@@ -94,7 +84,7 @@ let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor
 
         loop())
 
-let connectPropertyInspector (args: StartArgs) (agent: MailboxProcessor<PiInEvent>) =
+let connectPropertyInspector (args: StartArgs) (agent: MailboxProcessor<PiEvent>) =
     Utils.addDynamicStyles args.ApplicationInfo.colors
 
     let websocket = Utils.createWebSocket args.Port
@@ -107,18 +97,18 @@ let connectPropertyInspector (args: StartArgs) (agent: MailboxProcessor<PiInEven
                 uuid = args.UUID
             |}
 
-            agent.Post <| PiInEvent.Connected(args, replyAgent)
+            agent.Post <| PiEvent.Connected(args, replyAgent)
 
     websocket.onmessage <-
         fun messageEvent ->
-            let event = (Utils.parseJson messageEvent.data) :?> Event
+            let event = Utils.parseJson messageEvent.data :?> Event
             let payload = event.payload :?> ActionPayload
 
             let piEvent =
                 match event.event with
-                | "didReceiveSettings" -> Some <| PiInEvent.DidReceiveSettings(event, payload)
-                | "didReceiveGlobalSettings" -> Some <| PiInEvent.DidReceiveGlobalSettings(payload.settings)
-                | "sendToPropertyInspector" -> Some <| PiInEvent.SendToPropertyInspector(event)
+                | "didReceiveGlobalSettings" -> Some <| PiEvent.DidReceiveGlobalSettings payload.settings
+                | "sendToPropertyInspector" -> Some <| PiEvent.DidReceivePluginMessage event
+                | "didReceiveSettings" -> Some <| PiEvent.DidReceiveSettings(event, payload)
                 | _ ->
                     console.warn($"Unexpected event ({event.event}) received by Property Inspector", event)
                     None