+++ src/StreamDeck.SDK/PluginModel.fs
@@ -6,79 +6,84 @@ open Browser.Types
 open Dto
 
 [<RequireQualifiedAccess>]
-type PluginInEvent =
-    /// When `websocket.onopen` this event diliver initial state to plugin.
-    | Connected of startArgs: StartArgs * replyAgent: MailboxProcessor<PluginOutEvent>
-    /// Event received after calling the getSettings API to retrieve the persistent data stored for the action.
-    | DidReceiveSettings of event: Event * payload: ActionPayload
-    /// Event received after calling the getGlobalSettings API to retrieve the global persistent data.
-    | DidReceiveGlobalSettings of setting: obj
-    /// When the user presses a key, the plugin will receive the `keyDown` event.
-    | KeyDown of event: Event * payload: ActionPayload
-    /// When the user releases a key, the plugin will receive the `keyUp` event.
-    | KeyUp of event: Event * payload: ActionPayload
-    // When the user touches the display, the plugin will receive the touchTap event (SD+).
-    | TouchTap of event: Event * payload: TouchTapActionPayload
-    // When the user presses or releases the encoder, the plugin will receive the dialPress event (SD+).
-    | DialPress of event: Event * payload: DialPressActionPayload
-    // When the user rotates the encoder, the plugin will receive the dialRotate event (SD+).
-    | DialRotate of event: Event * payload: DialRotateActionPayload
-    /// When an instance of an action is displayed on the Stream Deck, for example when the hardware is first plugged in, or when a folder containing that action is entered, the plugin will receive a `willAppear` event.
-    | WillAppear of event: Event * payload: AppearanceActionPayload
-    /// When an instance of an action ceases to be displayed on Stream Deck, for example when switching profiles or folders, the plugin will receive a `willDisappear` event.
-    | WillDisappear of event: Event * payload: AppearanceActionPayload
-    /// When the user changes the title or title parameters, the plugin will receive a `titleParametersDidChange` event.
-    | TitleParametersDidChange of event: Event * payload: ActionTitlePayload
-    /// When a device is plugged to the computer, the plugin will receive a `deviceDidConnect` event.
-    | DeviceDidConnect of event: DeviceEvent
-    /// When a device is unplugged from the computer, the plugin will receive a `deviceDidDisconnect` event.
-    | DeviceDidDisconnect of event: DeviceEvent
-    /// When a monitored application is launched, the plugin will be notified and will receive the `applicationDidLaunch` event.
+type PluginEvent =
+    | Connected of startArgs: StartArgs * replyAgent: MailboxProcessor<PluginCommand>
+    /// Occurs when a monitored application is launched. Monitored applications can be defined in the manifest.json file via the Manifest.ApplicationsToMonitor property. See also ApplicationDidTerminate.
     | ApplicationDidLaunch of applicationId: string
-    /// When a monitored application is terminated, the plugin will be notified and will receive the `applicationDidTerminate` event.
+    /// Occurs when a monitored application terminates. Monitored applications can be defined in the manifest.json file via the Manifest.ApplicationsToMonitor property. See also ApplicationDidLaunch.
     | ApplicationDidTerminate of applicationId: string
-    /// When the computer is wake up, the plugin will be notified and will receive the `systemDidWakeUp` event.
-    | SystemDidWakeUp
-    /// Event received when the Property Inspector appears in the Stream Deck software user interface, for example when selecting a new instance.
+    /// Occurs when a Stream Deck device is connected. See also DeviceDidDisconnect.
+    | DeviceDidConnect of device: string * DeviceInfo: DeviceInfo
+    /// Occurs when a Stream Deck device is disconnected. See also DeviceDidConnect.
+    | DeviceDidDisconnect of device: string
+    /// Occurs when the user presses a dial (Stream Deck +). See also DialUp.
+    | DialDown of event: Event * payload: EncoderPayload
+    /// Occurs when the user rotates a dial (Stream Deck +).
+    | DialRotate of event: Event * payload: DialRotatePayload
+    /// Occurs when the user releases a pressed dial (Stream Deck +). See also DialDown.
+    | DialUp of event: Event * payload: EncoderPayload
+    /// Occurs when Stream Deck receives a deep-link message intended for the plugin. The message is re-routed to the plugin, and provided as part of the payload. One-way deep-link message can be routed to the plugin using the URL format streamdeck://plugins/message/<PLUGIN_UUID>/{MESSAGE}.
+    | DidReceiveDeepLink of url: string
+    /// Occurs when the plugin receives the global settings from the Stream Deck.
+    | DidReceiveGlobalSettings of setting: obj
+    /// Occurs when a payload was received from the UI.
+    | DidReceivePropertyInspectorMessage of event: Event
+    /// Occurs when the settings associated with an action instance are requested, or when the the settings were updated by the property inspector.
+    | DidReceiveSettings of event: Event * payload: ActionPayload
+    /// Occurs when the user presses a action down. See also KeyUp.
+    | KeyDown of event: Event * payload: ActionPayload
+    /// Occurs when the user releases a pressed action. See also KeyDown.
+    | KeyUp of event: Event * payload: ActionPayload
+    /// Occurs when the property inspector associated with the action becomes visible, i.e. the user selected an action in the Stream Deck application. See also PropertyInspectorDidDisappear.
     | PropertyInspectorDidAppear of event: Event
-    /// Event received when the Property Inspector for an instance is removed from the Stream Deck software user interface, for example when selecting a different instance.
+    /// Occurs when the property inspector associated with the action becomes invisible, i.e. the user unselected the action in the Stream Deck application. See also PropertyInspectorDidAppear.
     | PropertyInspectorDidDisappear of event: Event
-    /// Event received by the plugin when the Property Inspector uses the `sendToPlugin` event.
-    | SendToPlugin of event: Event
+    /// Occurs when the computer wakes up.
+    | SystemDidWakeUp
+    /// Occurs when the user updates an action's title settings in the Stream Deck application.
+    | TitleParametersDidChange of event: Event * payload: ActionTitlePayload
+    /// Occurs when the user taps the touchscreen (Stream Deck +).
+    | TouchTap of event: Event * payload: TouchTapActionPayload
+    /// Occurs when an action appears on the Stream Deck due to the user navigating to another page, profile, folder, etc. This also occurs during startup if the action is on the "front page". An action refers to all types of actions, e.g. keys, dials, touchscreens, pedals, etc.
+    | WillAppear of event: Event * payload: ActionPayload
+    /// Occurs when an action disappears from the Stream Deck due to the user navigating to another page, profile, folder, etc. An action refers to all types of actions, e.g. keys, dials, touchscreens, pedals, etc.
+    | WillDisappear of event: Event * payload: ActionPayload
 
-and [<RequireQualifiedAccess>] PluginOutEvent =
-    /// Save data persistently for the action's instance.
-    | SetSettings of context: string * payload: obj
-    /// Request the persistent data for the action's instance.
-    | GetSettings of context: string
-    /// Save data securely and globally for the plugin.
-    | SetGlobalSettings of payload: obj
-    /// Request the global persistent data.
+and [<RequireQualifiedAccess>] PluginCommand =
+    /// Gets the global settings associated with the plugin. Causes DidReceiveGlobalSettings to be emitted.
     | GetGlobalSettings
-    /// Open an URL in the default browser.
-    | OpenUrl of url: string
-    /// Write a debug log to the logs file.
+    /// Gets the settings associated with an instance of an action. Causes DidReceiveSettings to be emitted.
+    | GetSettings of context: string
+    /// Logs a message to the file-system.
     | LogMessage of message: string
-    /// Dynamically change the title of an instance of an action.
-    | SetTitle of context: string * payload: SetTitlePayload
-    /// Dynamically change the image displayed by an instance of an action.
-    | SetImage of context: string * payload: SetImagePayload
-    // Dynamically change properties of items on the Stream Deck + touch display.
+    /// Opens the URL in the user's default browser.
+    | OpenUrl of url: string
+    /// Sends a message to the property inspector.
+    | SendToPropertyInspector of context: string * payload: obj
+    /// Set's the feedback of an existing layout associated with an action instance.
     | SetFeedback of context: string * payload: Dictionary<string, obj>
-    // Dynamically change the current layout for the Stream Deck + touch display
+    /// Sets the layout associated with an action instance.
     | SetFeedbackLayout of context: string * payload: SetFeedbackLayoutPayload
-    /// Temporarily show an alert icon on the image displayed by an instance of an action.
-    | ShowAlert of context: string
-    /// Temporarily show an OK checkmark icon on the image displayed by an instance of an action.
-    | ShowOk of context: string
-    /// Change the state of the action's instance supporting multiple states.
+    /// Sets the global settings associated with the plugin.
+    | SetGlobalSettings of context: string * payload: obj
+    /// Sets the image associated with an action instance.
+    | SetImage of context: string * payload: SetImagePayload
+    /// Sets the settings associated with an instance of an action.
+    | SetSettings of context: string * payload: obj
+    /// Sets the settings associated with an instance of an action.
     | SetState of context: string * state: int
-    /// Switch to one of the preconfigured read-only profiles.
-    | SwitchToProfile of device: string * profileName: string
-    /// Send a payload to the Property Inspector.
-    | SendToPropertyInspector of context: string * action: string * payload: obj
+    /// Sets the title displayed for an instance of an action.
+    | SetTitle of context: string * payload: SetTitlePayload
+    /// Sets the trigger descriptions associated with an encoder action instance.
+    | SetTriggerDescription of context: string * payloaf: SetTriggerDescriptionPayload
+    /// Temporarily shows an alert (i.e. warning), in the form of an exclamation mark in a yellow triangle, on the action instance. Used to provide visual feedback when an action failed.
+    | ShowAlert of context: string
+    /// Temporarily shows an "OK" (i.e. success), in the form of a check-mark in a green circle, on the action instance. Used to provide visual feedback when an action successfully executed.
+    | ShowOk of context: string
+    /// Switches to the profile, as distributed by the plugin, on the specified device.
+    | SwitchToProfile of device: string * profileName: string * page: string
 
-let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor<PluginOutEvent> =
+let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor<PluginCommand> =
     let inPluginUUID = args.UUID
 
     MailboxProcessor.Start(fun inbox ->
@@ -91,103 +96,96 @@ let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor
                 console.log($"Plugin sent event %A{msg}", msg)
 
                 match msg with
-                | PluginOutEvent.SetSettings(context, payload) ->
-                    sendJson {|
-                        event = "setSettings"
-                        // An opaque value identifying the instance's action.
-                        context = context
-                        payload = payload
-                    |}
-                | PluginOutEvent.GetSettings context ->
-                    sendJson {|
-                        event = "getSettings"
-                        // An opaque value identifying the instance's action
-                        context = context
-                    |}
-                | PluginOutEvent.SetGlobalSettings(payload) ->
-                    sendJson {|
-                        event = "setGlobalSettings"
-                        // An opaque value identifying the plugin (inPluginUUID). This value is received during the Registration procedure.
-                        context = inPluginUUID
-                        payload = payload
-                    |}
-                | PluginOutEvent.GetGlobalSettings ->
+                | PluginCommand.GetGlobalSettings ->
                     sendJson {|
                         event = "getGlobalSettings"
-                        // An opaque value identifying the plugin (inPluginUUID). This value is received during the Registration procedure.
                         context = inPluginUUID
                     |}
-                | PluginOutEvent.OpenUrl url ->
+                | PluginCommand.GetSettings context ->
+                    sendJson {|
+                        event = "getSettings"
+                        context = context
+                    |}
+                | PluginCommand.OpenUrl url ->
                     sendJson {|
                         event = "openUrl"
                         payload = {| url = url |}
                     |}
-                | PluginOutEvent.LogMessage message ->
+                | PluginCommand.LogMessage message ->
                     sendJson {|
                         event = "logMessage"
                         payload = {| message = message |}
                     |}
-                | PluginOutEvent.SetTitle(context, payload) ->
+                | PluginCommand.SendToPropertyInspector(context, payload) ->
                     sendJson {|
-                        event = "setTitle"
-                        // An opaque value identifying the instance's action you want to modify.
+                        event = "sendToPropertyInspector"
                         context = context
                         payload = payload
                     |}
-                | PluginOutEvent.SetImage(context, payload) ->
-                    sendJson {|
-                        event = "setImage"
-                        // An opaque value identifying the instance's action you want to modify.
-                        context = context
-                        payload = payload
-                    |}
-                | PluginOutEvent.SetFeedback(context, payload) ->
+                | PluginCommand.SetFeedback(context, payload) ->
                     sendJson {|
                         event = "setFeedback"
-                        // A value to Identify the instance's action you want to modify.
                         context = context
                         payload = payload
                     |}
-                | PluginOutEvent.SetFeedbackLayout(context, payload) ->
+                | PluginCommand.SetFeedbackLayout(context, payload) ->
                     sendJson {|
                         event = "setFeedbackLayout"
                         context = context
                         payload = payload
                     |}
-                | PluginOutEvent.ShowAlert context ->
+                | PluginCommand.SetGlobalSettings(context, payload) ->
                     sendJson {|
-                        event = "showAlert"
-                        // An opaque value identifying the instance's action.
+                        event = "setGlobalSettings"
                         context = context
+                        payload = payload
                     |}
-                | PluginOutEvent.ShowOk context ->
+                | PluginCommand.SetImage(context, payload) ->
                     sendJson {|
-                        event = "showOk"
-                        // An opaque value identifying the instance's action.
+                        event = "setImage"
                         context = context
+                        payload = payload
                     |}
-                | PluginOutEvent.SetState(context, state) ->
+                | PluginCommand.SetSettings(context, payload) ->
+                    sendJson {|
+                        event = "setSettings"
+                        context = context
+                        payload = payload
+                    |}
+                | PluginCommand.SetState(context, state) ->
                     sendJson {|
                         event = "setState"
-                        // An opaque value identifying the instance's action.
                         context = context
                         payload = {| state = state |}
                     |}
-                | PluginOutEvent.SwitchToProfile(device, profileName) ->
+                | PluginCommand.SetTitle(context, payload) ->
                     sendJson {|
-                        event = "switchToProfile"
-                        // An opaque value identifying the plugin. This value should be set to the PluginUUID received during the registration procedure.
-                        context = inPluginUUID
-                        device = device
-                        payload = {| profile = profileName |}
-                    |}
-                | PluginOutEvent.SendToPropertyInspector(context, action, payload) ->
-                    sendJson {|
-                        action = action
-                        event = "sendToPropertyInspector"
-                        // An opaque value identifying the instance's action.
+                        event = "setTitle"
                         context = context
                         payload = payload
+                    |}
+                | PluginCommand.SetTriggerDescription(context, payload) ->
+                    sendJson {|
+                        event = "setTriggerDescription"
+                        context = context
+                        payload = payload
+                    |}
+                | PluginCommand.ShowAlert context ->
+                    sendJson {|
+                        event = "showAlert"
+                        context = context
+                    |}
+                | PluginCommand.ShowOk context ->
+                    sendJson {|
+                        event = "showOk"
+                        context = context
+                    |}
+                | PluginCommand.SwitchToProfile(device, profile, page) ->
+                    sendJson {|
+                        event = "switchToProfile"
+                        context = inPluginUUID
+                        device = device
+                        payload = {| profile = profile; page = page |}
                     |}
 
                 return! loop()
@@ -195,7 +193,7 @@ let createReplyAgent (args: StartArgs) (websocket: WebSocket) : MailboxProcessor
 
         loop())
 
-let connectPlugin (args: Dto.StartArgs) (agent: MailboxProcessor<PluginInEvent>) =
+let connectPlugin (args: Dto.StartArgs) (agent: MailboxProcessor<PluginEvent>) =
     let websocket = Utils.createWebSocket args.Port
     let replyAgent = createReplyAgent args websocket
 
@@ -206,7 +204,7 @@ let connectPlugin (args: Dto.StartArgs) (agent: MailboxProcessor<PluginInEvent>)
                 uuid = args.UUID
             |}
 
-            agent.Post <| PluginInEvent.Connected(args, replyAgent)
+            agent.Post <| PluginEvent.Connected(args, replyAgent)
 
     websocket.onmessage <-
         fun messageEvent ->
@@ -216,27 +214,46 @@ let connectPlugin (args: Dto.StartArgs) (agent: MailboxProcessor<PluginInEvent>)
 
             let pEvent =
                 match event.event with
-                | "didReceiveSettings" -> Some(PluginInEvent.DidReceiveSettings(event, payload))
-                | "didReceiveGlobalSettings" -> Some(PluginInEvent.DidReceiveGlobalSettings(payload.settings))
-                | "keyDown" -> Some(PluginInEvent.KeyDown(event, payload))
-                | "keyUp" -> Some(PluginInEvent.KeyUp(event, payload))
-                | "touchTap" -> Some(PluginInEvent.TouchTap(event, event.payload :?> TouchTapActionPayload))
-                | "dialPress" -> Some(PluginInEvent.DialPress(event, event.payload :?> DialPressActionPayload))
-                | "dialRotate" -> Some(PluginInEvent.DialRotate(event, event.payload :?> DialRotateActionPayload))
-                | "willAppear" -> Some(PluginInEvent.WillAppear(event, event.payload :?> AppearanceActionPayload))
-                | "willDisappear" -> Some(PluginInEvent.WillDisappear(event, event.payload :?> AppearanceActionPayload))
-                | "titleParametersDidChange" ->
-                    Some(PluginInEvent.TitleParametersDidChange(event, event.payload :?> ActionTitlePayload))
-                | "deviceDidConnect" -> Some(PluginInEvent.DeviceDidConnect(json :?> DeviceEvent))
-                | "deviceDidDisconnect" -> Some(PluginInEvent.DeviceDidDisconnect(json :?> DeviceEvent))
                 | "applicationDidLaunch" ->
-                    Some(PluginInEvent.ApplicationDidLaunch((json :?> ApplicationPayload).application))
+                    let p = event.payload :?> ApplicationPayload
+                    Some(PluginEvent.ApplicationDidLaunch p.application)
                 | "applicationDidTerminate" ->
-                    Some(PluginInEvent.ApplicationDidTerminate((json :?> ApplicationPayload).application))
-                | "systemDidWakeUp" -> Some(PluginInEvent.SystemDidWakeUp)
-                | "propertyInspectorDidAppear" -> Some(PluginInEvent.PropertyInspectorDidAppear(event))
-                | "propertyInspectorDidDisappear" -> Some(PluginInEvent.PropertyInspectorDidDisappear(event))
-                | "sendToPlugin" -> Some(PluginInEvent.SendToPlugin(event))
+                    let p = event.payload :?> ApplicationPayload
+                    Some(PluginEvent.ApplicationDidTerminate p.application)
+                | "deviceDidConnect" ->
+                    let e = json :?> DeviceEvent
+                    Some(PluginEvent.DeviceDidConnect(e.device, e.deviceInfo.Value))
+                | "deviceDidDisconnect" ->
+                    let e = json :?> DeviceEvent
+                    Some(PluginEvent.DeviceDidDisconnect e.device)
+                | "dialDown" ->
+                    let p = event.payload :?> EncoderPayload
+                    Some(PluginEvent.DialDown(event, p))
+                | "dialRotate" ->
+                    let p = event.payload :?> DialRotatePayload
+                    Some(PluginEvent.DialRotate(event, p))
+                | "dialUp" ->
+                    let p = event.payload :?> EncoderPayload
+                    Some(PluginEvent.DialDown(event, p))
+                | "didReceiveDeepLink" ->
+                    let p = event.payload :?> DidReceiveDeepLinkPayload
+                    Some(PluginEvent.DidReceiveDeepLink p.url)
+                | "didReceiveGlobalSettings" -> Some(PluginEvent.DidReceiveGlobalSettings payload.settings)
+                | "sendToPlugin" -> Some(PluginEvent.DidReceivePropertyInspectorMessage event)
+                | "didReceiveSettings" -> Some(PluginEvent.DidReceiveSettings(event, payload))
+                | "keyDown" -> Some(PluginEvent.KeyDown(event, payload))
+                | "keyUp" -> Some(PluginEvent.KeyUp(event, payload))
+                | "propertyInspectorDidAppear" -> Some(PluginEvent.PropertyInspectorDidAppear event)
+                | "propertyInspectorDidDisappear" -> Some(PluginEvent.PropertyInspectorDidDisappear event)
+                | "systemDidWakeUp" -> Some PluginEvent.SystemDidWakeUp
+                | "titleParametersDidChange" ->
+                    let p = event.payload :?> ActionTitlePayload
+                    Some(PluginEvent.TitleParametersDidChange(event, p))
+                | "touchTap" ->
+                    let p = event.payload :?> TouchTapActionPayload
+                    Some(PluginEvent.TouchTap(event, p))
+                | "willAppear" -> Some(PluginEvent.WillAppear(event, payload))
+                | "willDisappear" -> Some(PluginEvent.WillDisappear(event, payload))
                 | _ ->
                     console.warn($"Unexpected event ({event.event}) received by Plugin", event)
                     None