+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.6.0.0] - 2025-03-09
 
+- Update StreamDeck SDK (WebSocket API)
 - Migration to Elgato CLI
 - Dependencies update (.NET 9, Fable 4.24, Node 22, React 19)